requirements.txt
asgiref==3.8.1
astroid==2.15.8
autoflake==2.3.1
autopep8==2.1.0
awscli==1.40.2
boto3==1.38.3
botocore==1.38.3
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
colorama==0.4.6
crispy-bootstrap5==2024.10
distlib==0.3.9
dj-database-url==2.3.0
dj-stripe==2.9.0
Django==5.1.6
django-allauth==65.4.1
django-countries==7.6.1
django-crispy-forms==2.3
django-csp==4.0
django-environ==0.12.0
django-storages==1.14.6
django-vite==3.1.0
docutils==0.19
filelock==3.18.0
flake8==6.1.0
flake8-django==1.4
gunicorn==23.0.0
identify==2.6.9
idna==3.10
isort==5.13.2
jmespath==1.0.1
lazy-object-proxy==1.10.0
mccabe==0.7.0
nodeenv==1.9.1
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
pre-commit==3.7.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.30.0
whitenoise==6.9.0
wrapt==1.17.2
